fix: ViewModelProviderFactory, CalculatorViewModel 내부클래스에서 별도 클래스로 분리
