+++ calculator-app/src/main/java/camp/nextstep/edu/calculator/CalculatorViewModel.kt
@@ -1,12 +1,9 @@
 package camp.nextstep.edu.calculator
 
-import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import camp.nextstep.edu.calculator.data.repository.DataInjector
 import camp.nextstep.edu.calculator.domain.Calculator
 import camp.nextstep.edu.calculator.domain.Expression
 import camp.nextstep.edu.calculator.domain.Operator
@@ -86,17 +83,5 @@ class CalculatorViewModel(
 
     enum class EventType {
         SHOW_TOAST
-    }
-
-    class CalculatorViewModelFactory(private val context: Context) : ViewModelProvider.Factory {
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return if (modelClass.isAssignableFrom(CalculatorViewModel::class.java)) {
-                CalculatorViewModel(
-                    resultExpressionRepository = DataInjector.provideResultExpressionRepository(context)
-                ) as T
-            } else {
-                throw IllegalArgumentException()
-            }
-        }
     }
 }
